Refactor MVC
