+++ .gitignore
@@ -1,1 +1,6 @@
 /.vscode
+/node_modules
+/.parcel-cache
+/dist
+.prettierrc
+.DS_Store
