+++ package.json
@@ -1,0 +1,11 @@
+{
+  "name": "snappy-trails",
+  "version": "1.0.0",
+  "main": "index.js",
+  "repository": "https://github.com/LanesGood/snappy-trails.git",
+  "author": "LanesGood <Lane@developmentseed.org>",
+  "license": "MIT",
+  "devDependencies": {
+    "parcel": "^2.8.3"
+  }
+}
