+++ src/index.html
@@ -21,6 +21,7 @@
     integrity="sha256-WBkoXOwTeyKclOHuWtc+i2uENFpDZ9YPdf5Hf+D7ewM=" crossorigin=""></script>
   <script src="https://cdn.jsdelivr.net/npm/leaflet-providers@1.13.0/leaflet-providers.min.js"></script>
   <script defer src="./app.js"></script>
+  <script type="module" src="./js/controller.js"></script>
 
   <title>Snappy Trails</title>
 </head>
